Docker support
Currently aimed at simplifying development. Needs to replace the
webserver with something better than Django runserver for production.

Run docker-rundevel.sh to build and run a container on port 8000 with
a default supoeruser with pano/puppet as credentials. Will mount the
directory config/ from the git root at /etc/panopuppet in the container.